Implement GUI code for lineside equipment.
Add a bunch of UI controls for the various pieces of equipment that
exist: signals, divided into stems and heads; track sections, horizontal
and diagonal; points; and grey exit triangles. Introduce a clock source
that provides a single clock for all blinking objects in the display.
Add interaction support for calling and cancelling routes, setting
signals to flag-by, and keying points.