Turn Expired property into CheckExpired function.
This is helpful as it allows us to add a side effect: when CheckExpired
returns true, it also sets LocationCurrent to false. This means that if
a Train object is kept around after the train has left the region, it
will be considered to have a non-current location, which is more
correct.