Implement train describer berths.
Each berth is associated with a track circuit and shows the oldest train
which has advanced to that circuit (based on berthing rules which follow
called routes). Clicking on a berth calls up all trains in the berth
into the train detail panel, in order of time in the circuit from oldest
to youngest.