Add ID property to track circuit schema.
There are large holes (a few hundred) in the track circuit ID allocation
scheme. These holes are not large enough to make using an array to store
track circuits impractical, but they do make it undesirable to add
entries to the XAML files for all the nonexistent circuits (besides
which letting the array entries stay clear makes it easy to catch cases
of schema error where a nonexistent track circuit is used for
something).