Calculate straight-only divergence aspects.
Most signals, if showing an immediate divergence (red over proceed),
will calculate what that proceed on the lower head should be based on
the total distance the train can advance before stopping. A few signals,
however, calculate it based on the distance the train can advance before
stopping *or diverging*. Allow either behaviour to be selected on a
per-route basis.