Do not allow auto working when swinging points.
If points are swinging, the signal has not yet been cleared and a packet
will be sent to do that once the points all prove. Pushing the auto
working button sends a packet to set the signal to fleet. If the points
have not yet proved, the signal should not yet be cleared, so fleeting
it should not be allowed. Also, when the points do prove, the second
packet would overwrite the effect of the first and deactivate auto
working. Better to just not allow it until the points are proved.